Outsources implementation details for compute_sector and compute_area on BezierCurve
