+++ src/axom/primal/CMakeLists.txt
@@ -48,6 +48,7 @@ set( primal_headers
      operators/split.hpp
 
      operators/detail/clip_impl.hpp
+     operators/detail/compute_moments_impl.hpp
      operators/detail/intersect_bezier_impl.hpp
      operators/detail/intersect_bounding_box_impl.hpp
      operators/detail/intersect_impl.hpp
